fix(backend): bound the LLM timeout setting and its connect phase
Addresses the review backlog on #14206. Six fixes; the rest of the findings
got replies on the PR.

- Cap `llm_request_timeout_seconds` at `le=1500`. The field's own description
  mandated staying under the 1800s node cap but nothing enforced it, so
  `LLM_REQUEST_TIMEOUT_SECONDS=3600` validated cleanly and inverted the
  layering the PR is built on (QA reproduced this in a running executor).
  Literal rather than importing DEFAULT_BLOCK_EXECUTION_TIMEOUT_SECONDS —
  util must not import blocks.

- Bound the httpx connect phase at 5s via a new `request_timeout()` helper.
  A scalar timeout sets connect/read/write/pool alike, so the raise to 600s
  also stretched TCP connect 5x; both SDK defaults use connect=5.0. No
  generation happens during connect, so the PR's "no bytes until done"
  rationale does not cover it. Applied at the 6 real httpx sites; the
  `asyncio.wait_for` in call_provider keeps the scalar wall-clock deadline.

- Bound chat session-title generation (`copilot/service.py`) at 30s. It was
  the one production caller inheriting the shared default with no explicit
  timeout, so a 20-token title call got a 10-minute leash while holding a
  background task and a slot in the shared aux-client pool.

- Move the node-cap guard onto the schema bound instead of the live constant.
  The constant is env-derived, so the old assertion passed or failed on the
  runner's environment rather than on the code, and left a misconfigured
  deployment unguarded.

- Tests: pin the shipped 600s default (the previous assertion compared the
  constant to a fresh read of the same setting and so held for any value),
  reject 29/1501, assert both layers resolve to one deadline, assert connect
  is bounded separately, and cover the remediation hint — previously asserted
  nowhere despite being the only user-visible change.

- Drop stale/diff-narrating comments: `dream/llm.py` still documented a "120s
  default", and rename the dream regression test to name the 120s truncation
  threshold it actually asserts.

Verified by executing the synchronous test bodies directly (pytest needs
docker infra absent in this worktree) plus an interception check that
httpx.Timeout(connect=5.0, read=600) reaches the real SDK call site.
ruff/isort/black/pyright clean on all 9 files; docs-sync --check exits 0.

+++ autogpt_platform/backend/backend/blocks/llm.py
@@ -61,7 +61,7 @@ USER_ERROR_STATUS_CODES = (401, 403, 429)
 
 # Non-streaming calls emit no bytes until the completion is done, so this has
 # to cover the whole generation — a slow long answer is not a stalled socket.
-LLM_REQUEST_TIMEOUT_SECONDS: float = settings.config.llm_request_timeout_seconds
+LLM_REQUEST_TIMEOUT_SECONDS: int = settings.config.llm_request_timeout_seconds
 
 LLMProviderName = Literal[
     ProviderName.AIML_API,
